+++ Assets/Scripts/Core/Board.cs
@@ -140,10 +140,19 @@ public class Board
         fiftyRuleHalfClock = 0;
     }
 
+    public void BeforeLoadingPosition()
+    {
+        EnginePlayer.CancelSearch();
+
+        // while(Main.engine.IsSearching())
+        // {
+        //     // SOMETHING WRONG HERE: WHEN LOADING POSITION WHERE IT IS ENGINE'S TURN ****** AAA AEEAEAEAEA
+        // }
+    }
+
     public void AfterLoadingPosition()
     {
         currentLegalMoves = MoveGen.GenerateMoves(this);
-        EnginePlayer.CancelSearch();
     }
 
     public void Reset()