+++ Assets/Scripts/Graphics/PositionLoader.cs
@@ -73,6 +73,7 @@ public class PositionLoader : MonoBehaviour
 
     void LoadPositionFromFen(string fen)
     {
+        board.BeforeLoadingPosition();
         Setup();
 
         for (int i = 0; i < 12; i++)